apigateway/requirements.txt
django
django-core
djangorestframework
django-cors-headers
requests
PyJWT
python-dotenv
